File tidying, JSdoc writing, changelog adding...
