+++ src/syncer/partner_tcp.ts
@@ -1,6 +1,6 @@
 import { AsyncQueue, concat } from "../../deps.ts";
-import { TcpProvider } from "../lan/tcp_provider.ts";
-import { ITcpConn } from "../lan/types.ts";
+import { TcpProvider } from "../tcp/tcp_provider.ts";
+import { ITcpConn } from "../tcp/types.ts";
 import { NotSupportedError } from "../util/errors.ts";
 import { DecryptLengthDelimitStream, DecryptStream } from "./message_crypto.ts";
 import {
