+++ src/test/scenarios/scenarios.ts
@@ -33,8 +33,8 @@ import { getFreePort } from "https://deno.land/x/free_port@v1.2.0/mod.ts";
 import { Server } from "../../server/server.ts";
 import { IServerExtension } from "../../server/extensions/extension.ts";
 import { ExtensionSyncWeb } from "../../server/extensions/sync_web.ts";
-import { LANSession } from "../../lan/discovery_lan.ts";
-import { TcpProvider } from "../../lan/tcp_provider.ts";
+import { LANSession } from "../../discovery/discovery_lan.ts";
+import { TcpProvider } from "../../tcp/tcp_provider.ts";
 
 export const cryptoScenarios: Scenario<ICryptoDriver>[] = [
   ...universalCryptoDrivers,
